backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
passlib==1.7.4
Authlib==1.3.0
python-dotenv==1.0.1
